extract compare code out of query_helper.py
